Gets rid of problem with field 245*.a
In order to generate the hash identifiers for person and organisation documents, we need among others the title field of the record (MARC field 245*.a). Thus we can't output the identifier until the field 245*.a is processed. This is not a problem if the other required fields for a person/organisation documents are processed later (i.e. values from MARC fields 700 and above). It is however a problem if a person is generated from field 100: We can't wait to flush the data until processing field 245*.a without risking to mix up data belonging to different authors.
The solution is to promote the value of 245*.a as record identifier before the morph step. Thus the value of field 245*.a is always the first field being present in the stream.